api_test/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0
